requirements.txt
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
asyncmy==0.2.10
beanie==1.29.0
certifi==2025.4.26
cffi==1.17.1
ciso8601==2.3.2
click==8.2.0
cryptography==45.0.2
dnspython==2.7.0
fastapi==0.115.12
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
iso8601==2.1.0
lazy-model==0.2.0
motor==3.7.1
orjson==3.10.18
packaging==25.0
pluggy==1.6.0
pycparser==2.22
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
pymongo==4.13.0
pypika-tortoise==0.5.0
pytest==8.3.5
pytest-asyncio==1.0.0
python-dotenv==1.1.0
pytz==2025.2
redis==6.1.0
sniffio==1.3.1
starlette==0.46.2
toml==0.10.2
tortoise-orm==0.25.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzlocal==5.3.1
uvicorn==0.34.2
uvloop==0.21.0
